Add property-based tests to classes.
Property-based tests extensively test different kwargs combinations
and argument types.

This is the first of the 4 steps I propose for strict argument
validation with Pydantic. Arguments are currently not validated.

The proposed steps are as follows:
1. Create tests for core classes (to ensure changes are non-breaking)
2. Add type hinting to core classes
3. Add Pydantic models / validation for accepted arguments
   - this will include types and value ranges
   - this will also move out validation logic out of classes
4. Add GitHub pipeline to run tests on merges.

All of the above will make the code clearer, more modular and
more robust.